Fixed compiler warnings

+++ apc_cache.c
@@ -324,7 +324,7 @@ static void apc_cache_expunge(apc_cache_t* cache, size_t size TSRMLS_DC)
          * we run out of space.
          */
         CACHE_SAFE_LOCK(cache);
-        if (apc_sma_get_avail_mem() > (APCG(shm_size)/2)) {
+        if (apc_sma_get_avail_mem() > (size_t)(APCG(shm_size)/2)) {
             /* probably a queued up expunge, we don't need to do this */
             CACHE_SAFE_UNLOCK(cache);
             return;
@@ -353,7 +353,7 @@ clear_all:
          */
 
         CACHE_SAFE_LOCK(cache);
-        if (apc_sma_get_avail_mem() > (APCG(shm_size)/2)) {
+        if (apc_sma_get_avail_mem() > (size_t)(APCG(shm_size)/2)) {
             /* probably a queued up expunge, we don't need to do this */
             CACHE_SAFE_UNLOCK(cache);
             return;
